+++ mall-pms/pms-biz/src/main/java/com/youlai/mall/pms/controller/app/AppProductController.java
@@ -25,8 +25,8 @@ public class AppProductController {
 
     @ApiOperation(value = "列表分页", httpMethod = "GET")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "page", value = "页码", paramType = "query", dataType = "Integer"),
-            @ApiImplicitParam(name = "limit", value = "每页数量", paramType = "query", dataType = "Integer"),
+            @ApiImplicitParam(name = "page", value = "页码", paramType = "query", dataType = "Long"),
+            @ApiImplicitParam(name = "limit", value = "每页数量", paramType = "query", dataType = "Long"),
             @ApiImplicitParam(name = "name", value = "商品名称", paramType = "query", dataType = "String"),
             @ApiImplicitParam(name = "categoryId", value = "商品类目", paramType = "query", dataType = "Long")
     })
