+++ mall-pms/pms-biz/src/main/java/com/youlai/mall/pms/controller/app/AppSkuController.java
@@ -41,7 +41,7 @@ public class AppSkuController {
     @ApiOperation(value = "修改库存", httpMethod = "POST")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "id", value = "Sku ID", required = true, paramType = "path", dataType = "Long"),
-            @ApiImplicitParam(name = "num", value = "库存数量", required = true, paramType = "query", dataType = "Integer")
+            @ApiImplicitParam(name = "num", value = "库存数量", required = true, paramType = "query", dataType = "Long")
     })
     @PutMapping("/{id}/stock")
     public Result updateStock(@PathVariable Long id, @RequestParam Integer num) {
