+++ mall-ums/ums-biz/src/main/java/com/youlai/mall/ums/controller/app/AppAddressController.java
@@ -58,7 +58,7 @@ public class AppAddressController {
 
     @ApiOperation(value = "修改地址", httpMethod = "PUT")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "部门id", required = true, paramType = "path", dataType = "Integer"),
+            @ApiImplicitParam(name = "id", value = "部门id", required = true, paramType = "path", dataType = "Long"),
             @ApiImplicitParam(name = "address", value = "实体JSON对象", required = true, paramType = "body", dataType = "UmsAddress")
     })
     @PutMapping(value = "/{id}")
